Separate test diagnostics and timeout parsing
