+++ tools/testrunner/diagnostics_test.go
@@ -7,21 +7,6 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestParseTestTimeouts(t *testing.T) {
-	logInput, err := os.ReadFile("testdata/timeout-input.log")
-	require.NoError(t, err)
-	logOutput, err := os.ReadFile("testdata/timeout-output.log")
-	require.NoError(t, err)
-
-	stacktrace, tests := parseTestTimeouts(string(logInput))
-	require.Equal(t, []string{
-		"TestActivityApiResetClientTestSuite",
-		"TestNDCFuncTestSuite",
-		"TestActivityApiStateReplicationSuite",
-	}, tests)
-	require.Equal(t, string(logOutput), stacktrace)
-}
-
 func TestParseFailedTestsFromOutput(t *testing.T) {
 	t.Run("ExtractsFailedTestNames", func(t *testing.T) {
 		stdout := `
