+++ tools/testrunner/timeout_test.go
@@ -1,0 +1,29 @@
+package testrunner
+
+import (
+	"os"
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestParseTestTimeouts(t *testing.T) {
+	logInput, err := os.ReadFile("testdata/timeout-input.log")
+	require.NoError(t, err)
+	logOutput, err := os.ReadFile("testdata/timeout-output.log")
+	require.NoError(t, err)
+
+	stacktrace, tests := parseTestTimeouts(string(logInput))
+	require.Equal(t, []string{
+		"TestActivityApiResetClientTestSuite",
+		"TestNDCFuncTestSuite",
+		"TestActivityApiStateReplicationSuite",
+	}, tests)
+	require.Equal(t, string(logOutput), stacktrace)
+}
+
+func TestParseTestTimeoutsTruncatedAfterPanic(t *testing.T) {
+	stacktrace, tests := parseTestTimeouts("panic: test timed out after 1m")
+	require.Empty(t, tests)
+	require.Contains(t, stacktrace, "0 timed out test(s)")
+}
